Remove Multi-Target skill check of smart indirect switch.
This is because it is a headache to support multi-target (AI included). We'll need to intercept all multi-target attack sequence and remember which weapon is on which, so that we can check whether a target has mixed direct and indirect fire, and if it is mixed then we know we need to check multi-target skill. (or vice versa: check mix if lack the skill).  It may make more sense but this is more trouble than it is worth.